Breakup IModelTransformer class (#295)
This was a suggestion by copilot and I think aligned a little with one
of the comments in the service's teams issue about reducing size of some
of the files. I reviewed this work as copilot was making the changes,
and I feel its fine. However, I would like to get the perspective of the
service team. These are pretty dramatic changes, and I don't necessarily
know if I agree that its worthwhile.

Decompose IModelTransformer into focused internal classes

Extracts three internal classes from IModelTransformer (~3,500 → ~2,500
lines), each owning a single responsibility:

| Class | Responsibility | Lines |

|--------------------|-----------------------------------------------------------------------------|--------|
| SyncTypeResolver | Determines sync direction
(forward/reverse/not-sync) by inspecting ESAs | ~125 |
| ProvenanceManager | Manages scope ESA lifecycle, sync versioning,
provenance queries, iteration | ~1,000 |
| ElementResolver | Resolves target element via remap → FedGuid → Code
fallback | ~115 |

Dependency Design

All three classes receive IModelCloneContext via composition. The
dependency chain is:

 IModelCloneContext → SyncTypeResolver → ProvenanceManager
 IModelCloneContext → ElementResolver

ProvenanceManager computes _startingChangesetIndices from the context
internally. IModelTransformer retains thin delegate methods for
externally-used
APIs.

Breaking Changes

The following public/static members were removed from IModelTransformer:

 - determineSyncType() — moved to SyncTypeResolver.determineSyncType()
- noEsaSyncDirectionErrorMessage — moved to
SyncTypeResolver.noEsaSyncDirectionErrorMessage
- queryScopeExternalSourceAspect() — moved to
ProvenanceManager.queryScopeExternalSourceAspect()
- getProvenanceDb() / getProvenanceSourceDb() — use
_provenanceManager.getProvenanceDb() (now protected)
- getIsReverseSynchronization() / getIsForwardSynchronization() — use
_provenanceManager or _syncTypeResolver
- forEachTrackedElement() (static) — moved to
ProvenanceManager.forEachTrackedElement()
- initElementProvenanceOptions() (static) — moved to
ProvenanceManager.initElementProvenanceOptions()
- initRelationshipProvenanceOptions() (static) — moved to
ProvenanceManager.initRelationshipProvenanceOptions()

Retained delegates (used by imodel-transformations-service):
initElementProvenance, getSynchronizationVersion,
tryGetProvenanceScopeAspect,
initScopeProvenance, updateSynchronizationVersion — one-liner
pass-throughs on IModelTransformer.

_provenanceManager is now protected so subclasses can access the full
ProvenanceManager API directly.

Bug Fix: allowNoScopingESA removal

Removed the private allowNoScopingESA hack from SyncTypeResolver. Two
tests were using argsForProcessChanges for initial imports (invalid
workflow —
should use processAll), then setting allowNoScopingESA = true to
suppress the resulting error. The tests now correctly use processAll for
the initial
import.

Related fix: processAll now explicitly sets _sourceChangeDataState =
"has-changes" so that updateSynchronizationVersion writes the sync
version after a
full export. A noProvenance guard was added to
ProvenanceManager.updateSynchronizationVersion to prevent crashes when
no scope ESA exists.

Test Improvements

- Replaced brittle
transformer["_provenanceManager"]["_targetScopeProvenanceProps"] access
with ProvenanceManager.queryScopeExternalSourceAspect()
database queries
 - Initial import tests now use the correct processAll workflow

---------
