+++ packages/transformer/src/test/standalone/IModelTransformerHub.test.ts
@@ -85,6 +85,7 @@ import {
   ProcessChangesOptions,
   TransformerLoggerCategory,
 } from "../../imodel-transformer";
+import { ProvenanceManager } from "../../ProvenanceManager";
 import {
   CountingIModelImporter,
   HubWrappers,
@@ -248,10 +249,20 @@ describe("IModelTransformerHub", () => {
         targetBriefcase
       );
       await transformer1.process();
-      const scopingEsa1 = transformer1["_targetScopeProvenanceProps"];
-      const reverseSyncVersion1 =
-        scopingEsa1?.jsonProperties.reverseSyncVersion;
-      assert.isEmpty(reverseSyncVersion1);
+      const scopeEsaResult1 =
+        await ProvenanceManager.queryScopeExternalSourceAspect(
+          targetBriefcase,
+          {
+            id: undefined,
+            classFullName: ExternalSourceAspect.classFullName,
+            scope: { id: IModel.rootSubjectId },
+            kind: ExternalSourceAspect.Kind.Scope,
+            element: { id: IModel.rootSubjectId },
+            identifier: sourceBriefcase.iModelId,
+          }
+        );
+      const jsonProps1 = JSON.parse(scopeEsaResult1?.jsonProperties ?? "{}");
+      assert.isEmpty(jsonProps1.reverseSyncVersion ?? "");
       targetBriefcase.saveChanges();
       await targetBriefcase.pushChanges({
         description: "target changes for transformation 1",
@@ -279,15 +290,26 @@ describe("IModelTransformerHub", () => {
       await transformer2.updateSynchronizationVersion({
         initializeReverseSyncVersion: true,
       });
-      const scopingEsa2 = transformer2["_targetScopeProvenanceProps"];
-      const reverseSyncVersion2 =
-        scopingEsa2?.jsonProperties.reverseSyncVersion;
+      const scopeEsaResult2 =
+        await ProvenanceManager.queryScopeExternalSourceAspect(
+          targetBriefcase,
+          {
+            id: undefined,
+            classFullName: ExternalSourceAspect.classFullName,
+            scope: { id: IModel.rootSubjectId },
+            kind: ExternalSourceAspect.Kind.Scope,
+            element: { id: IModel.rootSubjectId },
+            identifier: sourceBriefcase.iModelId,
+          }
+        );
+      const jsonProps2 = JSON.parse(scopeEsaResult2?.jsonProperties ?? "{}");
+      const reverseSyncVersion2 = jsonProps2.reverseSyncVersion;
       assert.isNotEmpty(reverseSyncVersion2);
       const expectedReverseSyncVersion1 = `${targetBriefcase.changeset.id};${targetBriefcase.changeset.index}`;
       assert.equal(reverseSyncVersion2, expectedReverseSyncVersion1);
       // the recently pushed PendingReverseSync index should be equal to the latest target changeset index + 1
       const lastPendingReverseSyncIndex1 =
-        scopingEsa2?.jsonProperties.pendingReverseSyncChangesetIndices.pop();
+        jsonProps2.pendingReverseSyncChangesetIndices?.pop();
       assert.equal(
         lastPendingReverseSyncIndex1,
         (targetBriefcase.changeset.index ?? 0) + 1
@@ -398,17 +420,20 @@ describe("IModelTransformerHub", () => {
         assert.equal(sourceDbChanges.relationship.updateIds.size, 0);
         assert.equal(sourceDbChanges.relationship.deleteIds.size, 0);
 
+        // Initial import uses processAll to establish provenance
         const transformer = await TestIModelTransformer.create(
           sourceDb,
           targetDb,
-          {
-            argsForProcessChanges: {
-              startChangeset: { id: sourceDb.changeset.id },
-            },
-          }
+          {}
         );
-        transformer["_allowNoScopingESA"] = true;
         await transformer.process();
+        // // Verify processAll wrote the sync version so subsequent processChanges starts from correct index
+        // const syncVersionAfterProcessAll = await transformer["_provenanceManager"].getSynchronizationVersion();
+        // assert.equal(
+        //   syncVersionAfterProcessAll.index,
+        //   sourceDb.changeset.index,
+        //   "processAll should write sync version matching source changeset index"
+        // );
         transformer.dispose();
         targetDb.saveChanges();
         await targetDb.pushChanges({ accessToken, description: "Import #1" });
@@ -2211,13 +2236,22 @@ describe("IModelTransformerHub", () => {
       argsForProcessChanges: {},
     });
     await transformer.process();
-    let scopingEsa = transformer["_targetScopeProvenanceProps"];
-    expect(
-      scopingEsa?.jsonProperties.pendingSyncChangesetIndices.length
-    ).to.equal(1);
-    expect(scopingEsa?.jsonProperties.pendingSyncChangesetIndices[0]).to.equal(
-      4
+    targetDb.saveChanges();
+    // Query scope ESA from database instead of reaching into private internals
+    let scopeEsaResult = await ProvenanceManager.queryScopeExternalSourceAspect(
+      targetDb,
+      {
+        id: undefined,
+        classFullName: ExternalSourceAspect.classFullName,
+        scope: { id: IModel.rootSubjectId },
+        kind: ExternalSourceAspect.Kind.Scope,
+        element: { id: IModel.rootSubjectId },
+        identifier: sourceDb.iModelId,
+      }
     );
+    let scopeJsonProps = JSON.parse(scopeEsaResult?.jsonProperties ?? "{}");
+    expect(scopeJsonProps.pendingSyncChangesetIndices?.length).to.equal(1);
+    expect(scopeJsonProps.pendingSyncChangesetIndices[0]).to.equal(4);
     transformer.dispose();
 
     // Open sourceDb not at tip
@@ -2239,10 +2273,20 @@ describe("IModelTransformerHub", () => {
       argsForProcessChanges: {},
     });
     await transformer.process();
-    scopingEsa = transformer["_targetScopeProvenanceProps"];
-    expect(
-      scopingEsa?.jsonProperties.pendingSyncChangesetIndices
-    ).to.deep.equal([4]);
+    targetDb.saveChanges();
+    scopeEsaResult = await ProvenanceManager.queryScopeExternalSourceAspect(
+      targetDb,
+      {
+        id: undefined,
+        classFullName: ExternalSourceAspect.classFullName,
+        scope: { id: IModel.rootSubjectId },
+        kind: ExternalSourceAspect.Kind.Scope,
+        element: { id: IModel.rootSubjectId },
+        identifier: sourceDbNotAtTip.iModelId,
+      }
+    );
+    scopeJsonProps = JSON.parse(scopeEsaResult?.jsonProperties ?? "{}");
+    expect(scopeJsonProps.pendingSyncChangesetIndices).to.deep.equal([4]);
     transformer.dispose();
   });
 
@@ -3377,11 +3421,10 @@ describe("IModelTransformerHub", () => {
         sourceDb,
         DetachedExportElementAspectsStrategy
       );
+      // First transformation uses processAll (no argsForProcessChanges) to establish provenance
       const transformer = new IModelTransformer(exporter, targetDb, {
         includeSourceProvenance: true,
-        argsForProcessChanges: {},
       });
-      transformer["_allowNoScopingESA"] = true;
 
       // run first transformation
       await transformer.process();
